Add doc for workaround to jenkins logs bug
Yesterday, Jenkins filled the /var partition with lots of logs. This is
this bug reported here:

  https://issues.jenkins-ci.org/browse/JENKINS-29490

The workaround is to disable support for multicast DNS at Java level in
Jenkins environment file.

The changes to this file are now given as a patch for the file provided
by the package.